backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
pandas==2.3.2
numpy==2.3.3
scikit-learn==1.7.2
scipy==1.16.2
PyJWT==2.9.0
werkzeug==3.1.3
PyPDF2==3.0.1
python-docx==1.1.0

